% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisprVerse.R
\docType{package}
\name{crisprVerse-package}
\alias{crisprVerse}
\alias{crisprVerse-package}
\title{crisprVerse: Easily install and load the crisprVerse ecosystem for CRISPR gRNA design}
\description{
The crisprVerse is a modular ecosystem of R packages developed for the design and manipulation of CRISPR guide RNAs (gRNAs). All packages share a common language and design principles. This package is designed to make it easy to install and load the crisprVerse packages in a single step. To learn more about the crisprVerse, visit \url{https://www.github.com/crisprVerse}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crisprVerse/crisprVerse}
  \item Report bugs at \url{https://github.com/crisprVerse/crisprVerse/issues}
}

}
\author{
\strong{Maintainer}: Jean-Philippe Fortin \email{fortin946@gmail.com}

}
\keyword{internal}
