% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{crisprVerse_deps}
\alias{crisprVerse_deps}
\title{List all crisprVerse dependencies}
\usage{
crisprVerse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
crisprVerse packages.}
}
\description{
List all crisprVerse dependencies
}
