% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crisprVerse_packages}
\alias{crisprVerse_packages}
\title{List all packages in the crisprVerse}
\usage{
crisprVerse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include crisprVerse in the list?}
}
\value{
A character vector of packages in the crisprVerse
}
\description{
List all packages in the crisprVerse
}
\examples{
crisprVerse_packages()
}
