% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{crisprVerse_update}
\alias{crisprVerse_update}
\title{Update crisprVerse packages}
\usage{
crisprVerse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
crisprVerse packages.}
}
\description{
This will check to see if all crisprVerse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
if (interactive()){
    crisprVerse_update()
}
}
