% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisprVerse.R
\docType{package}
\name{crisprVerse-package}
\alias{crisprVerse}
\alias{crisprVerse-package}
\title{crisprVerse: Collection of R packages for CRISPR gRNA design}
\description{
Something
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/crisprVerse/crisprVerse}
  \item Report bugs at \url{https://github.com/crisprVerse/crisprVerse/issues}
}

}
\author{
\strong{Maintainer}: Jean-Philippe Fortin \email{fortin946@gmail.com}

}
\keyword{internal}
